Added GE Blue and Red Colors in config and rename hero component
